Abstract some PhoneNumber structure in common code
